BED-6521 more renames
