+++ packages/javascript/bh-shared-ui/src/hooks/usePZParams/index.ts
@@ -13,7 +13,6 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
-import { useZoneQueryParams } from './useZoneQueryParams';
 
 export * from './usePZPathParams';
-export { useZoneQueryParams };
+export * from './usePZQueryParams';
